removed the thing I thought was smart but turned out to be stupid
Removed coerce value and PropertyChanged delegate on Color and Opacity properties.
Initially just the Color property was taken into account in the renderers, it was built up from Opacity property changed.
That was a mistake since it can sometimes breaks the notify property changed mechanism (for example if you set opacity to 1).